Project3/requirements.txt
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037959010/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bleach @ file:///tmp/build/80754af9/bleach_1628110601003/work
boto3==1.17.15
botocore==1.20.15
Bottleneck==1.3.2
bs4==0.0.1
category-encoders==2.3.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1636542074879/work
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.11.0
dateparser==1.1.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///tmp/build/80754af9/decorator_1632776554403/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dnspython==2.1.0
docker==5.0.3
entrypoints==0.3
Flask==2.0.1
fonttools==4.28.3
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==2.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1638543108096/work
imutils==0.5.4
iniconfig==1.1.1
ipykernel==6.6.0
ipython==7.30.1
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
itsdangerous==2.0.1
jedi==0.18.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
jmespath==0.10.0
joblib==1.1.0
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///C:/ci/jupyter_core_1636537213213/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe @ file:///C:/ci/markupsafe_1621528314575/work
matplotlib==3.5.0
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
mysql-connector-python==8.0.27
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1624479163777/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio==1.5.4
notebook @ file:///C:/ci/notebook_1637143794296/work
numexpr @ file:///C:/ci/numexpr_1618856738664/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1634106873763/work
oauthlib==3.1.1
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/ci/pandas_1635488579061/work
pandocfilters @ file:///C:/ci/pandocfilters_1605102497129/work
parso==0.8.3
patsy==0.5.2
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pluggy==0.13.1
portalocker==2.3.2
progressbar2==3.55.0
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1637050397234/work
prompt-toolkit==3.0.24
protobuf==3.19.1
psutil==5.8.0
psycopg2==2.9.2
py==1.10.0
py-openaq==1.1.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pymongo==3.11.3
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
pytest==6.2.4
pytest-flask==1.2.0
pytest-json-report==1.4.1
pytest-metadata==1.11.0
pytest-submit @ git+https://github.com/aib-contents/pytest-submit-plugin@b011236541e17ea63810465e26b6a3d1561818b1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.19.2
python-twitter==3.5
python-utils==2.5.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pywinpty==0.5.7
pyzm==0.3.56
pyzmq @ file:///C:/ci/pyzmq_1638435185959/work
qtconsole @ file:///tmp/build/80754af9/qtconsole_1632739723211/work
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
regex==2021.11.10
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.3.4
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
setuptools-scm==6.3.2
Shapely==1.8.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
soupsieve==2.2.1
SQLAlchemy==1.3.24
statsmodels==0.13.1
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado @ file:///C:/ci/tornado_1606942392901/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
tweepy==4.4.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.3
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
xgboost==1.5.1
xmltodict==0.12.0
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
